// apps/web/lib/generated/prisma/models/TeamMember.ts
export type TeamMemberModel = runtime.Types.Result.DefaultSelection<Prisma.$TeamMemberPayload>
// ...
export type $TeamMemberPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "TeamMember"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
    team: Prisma.$TeamPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    teamId: string
    role: $Enums.UserRole
    createdAt: Date
  }, ExtArgs["result"]["teamMember"]>
  composites: {}
}